Extract AbstractSerialization
