+++ core/src/main/java/io/lionweb/lioncore/java/serialization/BulkImportElement.java
@@ -4,7 +4,7 @@ import io.lionweb.lioncore.java.model.Node;
 import io.lionweb.lioncore.java.serialization.data.MetaPointer;
 
 public class BulkImportElement {
-    public String container;
-    public MetaPointer containment;
-    public Node tree;
+  public String container;
+  public MetaPointer containment;
+  public Node tree;
 }
